requirements.txt
# Core dependencies
numpy>=1.24.0
PyYAML>=6.0
requests>=2.31.0

# Speech recognition
whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1 